reset conn->rx_task in scsi_cmnd_start()
In scsi_cmnd_start() it looks like conn->rx_task is intended to be set only
during the call to _stage1_done, but there was no code to reset it.  I think it
probably doesn't actually matter in execution, but it was a little confusing to
wonder and try to understand why it wasn't reset.

Signed-off-by: David Butterfield <dab21774@gmail.com>