Add an XXX comment about two possible leak fixes.
I made a fix to a memory leak in my copy of SCST, and someone else fixed the
same leak in the trunk, which I noticed when I updated my workspace to current
trunk.  I don't know which is better; the XXX is a reminder to figure that out.

The fix that was put into trunk calls scst_acn_sysfs_del(), whose name makes me
wonder whether it will handle the memory free in a PROCFS build, which is the
type of build where I discovered the leak.

Signed-off-by: David Butterfield <dab21774@gmail.com>